rename: 진료기록 관련 화면 경로 수정
