+++ app/src/main/java/com/project/meongcare/medicalrecord/view/MedicalRecordInfoFragment.kt
@@ -1,4 +1,4 @@
-package com.project.meongcare
+package com.project.meongcare.medicalRecord.view
 
 import android.os.Bundle
 import android.view.LayoutInflater
