+++ app/src/main/res/layout/fragment_medical_record_info.xml
@@ -11,7 +11,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".MedicalRecordInfoFragment">
+        tools:context=".medicalRecord.view.MedicalRecordInfoFragment">
 
         <androidx.constraintlayout.widget.Guideline
             android:id="@+id/guideline_medicalrecordinfo_top"
